Ficus rebase: Course Dashboard > Remove Student Map
Removes the Course Student Map functionality, while leaving the Global Student Map intact.
Moves the view tests into test_global_views, so that the Global Student Map remains covered by tests.
