+++ course_dashboard/templates/course_dashboard/base_course.html
@@ -13,7 +13,6 @@
 
 <%block name="navigation">
     <ul id="container-navigation" class="nav nav-pills">
-        <li role="presentation" ${"class='active'" if active_tab == 'student_map' else '' }><a href="${reverse('course-dashboard:student-map', kwargs={'course_id': course_id})}">${_("Student map")}</a></li>
         <li role="presentation" ${"class='active'" if active_tab == 'forum_activity' else '' }><a href="${reverse('course-dashboard:forum-activity', kwargs={'course_id': course_id})}">${_("Forum activity")}</a></li>
         <li role="presentation" ${"class='active'" if active_tab == 'wiki_activity' else '' }><a href="${reverse('course-dashboard:wiki-activity', kwargs={'course_id': course_id})}">${_("Wiki activity")}</a></li>
         <li role="presentation" ${"class='active'" if active_tab == 'answers_distribution' else '' } ><a href="${reverse('course-dashboard:problem-stats:index', kwargs={'course_id': course_id})}">${_("Answers distribution")}</a></li>
