+++ course_dashboard/templates/course_dashboard/student-map-content.html
@@ -36,11 +36,7 @@
 
 <div class="row">
     <div class="col-md-12">
-        % if course_id:
-        <a class="pull-right" href="${reverse('course-dashboard:student-map', kwargs={'course_id': course_id})}?format=csv">${_("Download raw CSV data")}</a>
-        % else:
         <a class="pull-right" href="${reverse('course-dashboard-global:student-map')}?format=csv">${_("Download raw CSV data")}</a>
-        % endif
     </div>
 </div>
 <div class="row">
