--- course_dashboard/templates/course_dashboard/student-map.html
@@ -1,10 +1,0 @@
-## mako
-<%inherit file="base_course.html" />
-
-<%block name="content">
-    <%include file="student-map-content.html"/>
-</%block>
-
-<%block name="extra_footer">
-    <%include file="student-map-footer.html"/>
-</%block>
