+++ course_dashboard/tests/test_views.py
@@ -10,51 +10,6 @@ from student.tests.factories import UserFactory
 from .base import BaseCourseDashboardTestCase
 
 
-class StudentMapTestCase(BaseCourseDashboardTestCase):
-
-    def get_course_student_map(self, course, response_format=None):
-        return self.get_response("course-dashboard:student-map", course, response_format=response_format)
-
-    def get_student_map(self, course_id, response_format=None):
-        return self.get_course_id_response("course-dashboard:student-map", course_id, response_format=response_format)
-
-    def test_student_map(self):
-        response = self.get_course_student_map(self.course)
-
-        self.assertEqual(200, response.status_code)
-
-    def test_student_map_for_non_existing_course_raises_404(self):
-        response = self.get_student_map("not/really/there")
-        self.assertEqual(404, response.status_code)
-
-    def test_non_empty_student_map(self):
-        self.enroll_student(self.course, user__profile__country='FR')
-        response = self.get_course_student_map(self.course)
-        self.assertEqual(200, response.status_code)
-        self.assertIn("France", response.content)
-
-    def test_student_has_no_access(self):
-        student = UserFactory.create()
-        self.client.login(username=student.username, password="test")
-        response = self.get_course_student_map(self.course)
-        self.assertEqual(404, response.status_code)
-
-    def test_staff_member_has_access(self):
-        student = UserFactory.create(is_staff=True)
-        self.client.login(username=student.username, password="test")
-        response = self.get_course_student_map(self.course)
-        self.assertEqual(200, response.status_code)
-
-    def test_csv_response(self):
-        self.enroll_student(self.course, user__profile__country='FR')
-
-        response = self.get_course_student_map(self.course, response_format="csv")
-        rows = self.get_csv_response_rows(response)
-
-        self.assertEqual(200, response.status_code)
-        self.assertEqual(2, len(rows))
-        self.assertEqual([_("France"), "1"], rows[1])
-
 class CertificateStatsTestCase(BaseCourseDashboardTestCase):
     def setUp(self):
         super(CertificateStatsTestCase, self).setUp()
